Rename dense_bn_fuse -> bn_fuse
