+++ hls4ml/model/optimizer/passes/bn_fuse.py
